backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
pydantic==2.6.1
pydantic[email]
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart
celery==5.3.6
websockets==12.0
alembic==1.13.1
huggingface-hub==0.24.6
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy>=1.26,<2

transformers==4.48.3
python-jose[cryptography]==3.3.0
authlib==1.3.0
httpx==0.27.0
